refactor: adicao do BuscaService
