+++ back-end/src/main/java/dev/gui/processo_sergipetec/service/BuscaService.java
@@ -1,0 +1,27 @@
+package dev.gui.processo_sergipetec.service;
+
+import dev.gui.processo_sergipetec.model.VeiculoModel;
+import dev.gui.processo_sergipetec.repository.BuscaRepository;
+
+import java.sql.SQLException;
+import java.util.List;
+
+public class BuscaService {
+    private final BuscaRepository buscaRepository;
+
+    public BuscaService(BuscaRepository buscaRepository) {
+        this.buscaRepository = buscaRepository;
+    }
+
+    public List<VeiculoModel> buscarTodosVeiculos() throws SQLException {
+        return buscaRepository.listarVeiculos();
+    }
+
+    public VeiculoModel buscarVeiculoPorId(int id)  {
+        return buscaRepository.listarVeiculoPorId(id);
+    }
+
+    public List<VeiculoModel> buscaPersonalizada(String tipo, String modelo, String cor, Integer ano) throws SQLException {
+        return buscaRepository.consultarVeiculos(tipo, modelo, cor, ano);
+    }
+}
